split STRING to its own files; no code changes other than that
