+++ daemon/main.c
@@ -1004,6 +1004,7 @@ int main(int argc, char **argv) {
                             if(test_dbengine()) return 1;
 #endif
                             if(test_sqlite()) return 1;
+                            if(string_unittest(10000)) return 1;
                             if (dictionary_unittest(10000))
                                 return 1;
                             if (rrdlabels_unittest())
@@ -1027,6 +1028,9 @@ int main(int argc, char **argv) {
                         }
                         else if(strcmp(optarg, "dicttest") == 0) {
                             return dictionary_unittest(10000);
+                        }
+                        else if(strcmp(optarg, "stringtest") == 0) {
+                            return string_unittest(10000);
                         }
                         else if(strcmp(optarg, "rrdlabelstest") == 0) {
                             return rrdlabels_unittest();
