+++ libnetdata/dictionary/dictionary.h
@@ -304,27 +304,6 @@ extern struct dictionary_stats dictionary_stats_category_other;
 extern int dictionary_unittest(size_t entries);
 
 // ----------------------------------------------------------------------------
-// STRING implementation
-
-typedef struct netdata_string STRING;
-extern STRING *string_strdupz(const char *str);
-extern STRING *string_dup(STRING *string);
-extern void string_freez(STRING *string);
-extern size_t string_strlen(STRING *string);
-extern const char *string2str(STRING *string) NEVERNULL;
-
-// keep common prefix/suffix and replace everything else with [x]
-extern STRING *string_2way_merge(STRING *a, STRING *b);
-
-static inline int string_cmp(STRING *s1, STRING *s2) {
-    // STRINGs are deduplicated, so the same strings have the same pointer
-    // when they differ, we do the typical strcmp() comparison
-    return (s1 == s2)?0:strcmp(string2str(s1), string2str(s2));
-}
-
-extern void string_statistics(size_t *inserts, size_t *deletes, size_t *searches, size_t *entries, size_t *references, size_t *memory, size_t *duplications, size_t *releases);
-
-// ----------------------------------------------------------------------------
 // THREAD CACHE
 
 extern void *thread_cache_entry_get_or_set(void *key,
